Actually post-install delete project's `.eslintrc`.
Actually delete the dependent project's `.eslintrc` file in the
post-installation hook, so we are made sure the rules are updated from
the culture repository on a package update. Unfortunately there doesn't
seem to be a more elegant way to do this (at least with Node.js 0.10's
npm, since none of the `node_config_*` environment variables point to
the dependent package directory.